Add files via upload
So I tried following along on the videos that were uploaded to Canvas. I have been doing this for the past two hours and can't figure out why I cannot upload directly from terminal to GitHub. I think a major issue for me was that I missed Thursday, which is when I believe the class went into greater detail for the assignment. I am hoping that doing this is satisfactory enough for now and in the future, I will do it properly.